requirements.txt
setuptools
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
pyarrow
jinja2
kaleido
tabulate
dill
PyYAML
fastapi
uvicorn
python-multipart
statsmodels
lifelines
shap
sdv
mlflow
lightgbm
xgboost
